backend/requirements.txt
click
requests
types-requests
psycopg
psycopg[binary]
tqdm
types-tqdm
dotenv
statsmodels
google-cloud-logging
ruff
mypy
